% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_anova.R
\name{bake.step_anova}
\alias{bake.step_anova}
\title{the general bake function used for step_anova}
\usage{
\method{bake}{step_anova}(object, new_data, ...)
}
\arguments{
\item{object}{a output of prep}

\item{new_data}{a new data}

\item{...}{other arguments}
}
\value{

}
\description{
the general bake function used for step_anova
}
