% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_mrmr.R
\name{bake.step_mrmr}
\alias{bake.step_mrmr}
\title{The general bake function for step_mrmr}
\usage{
\method{bake}{step_mrmr}(object, new_data, ...)
}
\arguments{
\item{object}{output of prep}

\item{new_data}{new data}

\item{...}{other arguments}
}
\value{

}
\description{
The general bake function for step_mrmr
}
