% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_anova.R
\name{prep.step_anova}
\alias{prep.step_anova}
\title{the general prep function used for step_anova}
\usage{
\method{prep}{step_anova}(x, training, info = NULL)
}
\arguments{
\item{x}{a recipe object}

\item{training}{the data used in recipe}

\item{info}{...}
}
\value{

}
\description{
the general prep function used for step_anova
}
