% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_mrmr.R
\name{prep.step_mrmr}
\alias{prep.step_mrmr}
\title{The general prep function used for step_mrmr}
\usage{
\method{prep}{step_mrmr}(x, training, info = NULL)
}
\arguments{
\item{x}{a recipe object}

\item{training}{the training data}

\item{info}{...}
}
\value{

}
\description{
The general prep function used for step_mrmr
}
