% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_anova.R
\name{step_anova}
\alias{step_anova}
\title{A step_* method used for recipe to filter the features}
\usage{
step_anova(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  ref_dist = NA,
  options = list(p_thresh = 0.05),
  skip = F,
  id = rand_id("anova")
)
}
\arguments{
\item{recipe}{a recipe object}

\item{...}{the selected features}

\item{role}{see recipe description}

\item{trained}{see recipe description}

\item{ref_dist}{the selected feature names, you didn't need to assign}

\item{options}{the options to run step_anova}

\item{skip}{see recipe description}

\item{id}{the id of this step}
}
\value{
a recipe object
}
\description{
A step_* method used for recipe to filter the features
}
\examples{

data(cells)
cells <- mutate(cells) \%>\% select(-case)
colnames(cells)[1] <- 'Label'
a_rec <- recipe(Label~., data = cells) \%>\%
step_anova(all_predictors())
}
