% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/step_mrmr.R
\name{step_mrmr}
\alias{step_mrmr}
\title{a step_* method used in recipe to filter the features using MRMR method}
\usage{
step_mrmr(
  recipe,
  ...,
  role = NA,
  trained = FALSE,
  ref_dist = NA,
  options = list(num_feature = 30),
  skip = F,
  id = rand_id("mrmr")
)
}
\arguments{
\item{recipe}{a recipe object}

\item{...}{the selected features}

\item{role}{see recipe description}

\item{trained}{see recipe description}

\item{ref_dist}{the final selected feature names}

\item{options}{containing the number of features to remain}

\item{skip}{see recipe description}

\item{id}{see recipe description}
}
\value{
a recipe object
}
\description{
a step_* method used in recipe to filter the features using MRMR method
}
\examples{

data(cells)
cells <- mutate(cells) \%>\% select(-case)
colnames(cells)[1] <- 'Label'
a_rec <- recipe(Label~., data = cells) \%>\%
step_mrmr(all_predictors(), options = list(num_feature = 30))
}
